+++ R/fabric_livy_session.R
@@ -293,7 +293,9 @@ FabricLivySession <- R6::R6Class(
       fabric_livy_check_flag(verbose, "verbose")
       self$high_concurrency <- high_concurrency
       self$verbose <- verbose
-      private$credential <- credential
+      credential_reference <- fabric_livy_credential_reference(credential)
+      private$credential_ref <- credential_reference$reference
+      private$credential_key <- credential_reference$key
       type <- if (high_concurrency) {
         "highConcurrencySessions"
       } else {
@@ -352,7 +354,7 @@ FabricLivySession <- R6::R6Class(
         private$update(fabric_livy_json(
           "GET",
           self$url,
-          private$credential,
+          fabric_livy_handle_credential(private$credential_ref),
           deadline = deadline
         ))
       }
@@ -439,7 +441,7 @@ FabricLivySession <- R6::R6Class(
       response <- fabric_livy_json(
         "POST",
         endpoint,
-        private$credential,
+        fabric_livy_handle_credential(private$credential_ref),
         payload = Filter(
           Negate(is.null),
           list(
@@ -454,7 +456,7 @@ FabricLivySession <- R6::R6Class(
         session = self,
         response = response,
         url = paste0(endpoint, "/", response$id),
-        credential = private$credential,
+        credential = fabric_livy_handle_credential(private$credential_ref),
         verbose = self$verbose
       )
     },
@@ -513,7 +515,7 @@ FabricLivySession <- R6::R6Class(
         page <- fabric_livy_json(
           "GET",
           request$url,
-          private$credential
+          fabric_livy_handle_credential(private$credential_ref)
         )
         page_statements <- page$statements %||% list()
         if (!is.list(page_statements)) {
@@ -574,7 +576,7 @@ FabricLivySession <- R6::R6Class(
       fabric_livy_ok(
         "POST",
         paste0(self$url, "/reset-timeout"),
-        private$credential,
+        fabric_livy_handle_credential(private$credential_ref),
         idempotent = FALSE
       )
       invisible(self)
@@ -591,7 +593,7 @@ FabricLivySession <- R6::R6Class(
       fabric_livy_ok(
         "DELETE",
         self$url,
-        private$credential,
+        fabric_livy_handle_credential(private$credential_ref),
         idempotent = TRUE,
         accepted_status = 404L,
         deadline = deadline
@@ -602,7 +604,8 @@ FabricLivySession <- R6::R6Class(
     }
   ),
   private = list(
-    credential = NULL,
+    credential_ref = NULL,
+    credential_key = NULL,
     collection_url = NULL,
 
     # Check that the session is still usable before a public method sends a
@@ -712,7 +715,9 @@ FabricLivyStatement <- R6::R6Class(
       self$started_local <- Sys.time()
       self$verbose <- verbose
       private$session <- session
-      private$credential <- credential
+      credential_reference <- fabric_livy_credential_reference(credential)
+      private$credential_ref <- credential_reference$reference
+      private$credential_key <- credential_reference$key
       invisible(self)
     },
 
@@ -741,7 +746,7 @@ FabricLivyStatement <- R6::R6Class(
         self$response <- fabric_livy_json(
           "GET",
           self$url,
-          private$credential,
+          fabric_livy_handle_credential(private$credential_ref),
           deadline = deadline
         )
         self$state <- self$response$state %||% self$state
@@ -846,7 +851,7 @@ FabricLivyStatement <- R6::R6Class(
       response <- fabric_livy_json(
         "POST",
         paste0(self$url, "/cancel"),
-        private$credential,
+        fabric_livy_handle_credential(private$credential_ref),
         idempotent = FALSE
       )
       invisible(response)
@@ -854,7 +859,8 @@ FabricLivyStatement <- R6::R6Class(
   ),
   private = list(
     session = NULL,
-    credential = NULL
+    credential_ref = NULL,
+    credential_key = NULL
   ),
   cloneable = FALSE
 )